completed loading all tasks on load and loading each submitted task
