+++ public/scripts/tasks.js
@@ -15,6 +15,7 @@ $(() => {
      })
       .then((res) => {
         console.log('Getting back...', res);
+        loadTasks(true);
       })
 
   })
@@ -39,6 +40,72 @@ $(() => {
       ? taskForm.addClass('focus')
       : taskForm.removeClass('focus');
   });
+
+  /* loadTasks function */
+  // get tasks from database
+  const loadTasks = (onlyLoadLatest = false) => {
+    $.ajax('/tasks/api', { method: 'GET' })
+      .then((data) => {
+        // console.log('DATA', data);
+
+        if (onlyLoadLatest) {
+          renderTasks([data[data.length - 1]]);
+        } else {
+          renderTasks(data);
+        }
+      });
+  };
+  loadTasks();
+
+  // render multiple tasks
+  // tasks = [ {task, category_id} ]
+  const renderTasks = (tasks) => {
+    // initialize empty array to store rendered tasks
+    const renderedTasks_movies = [];
+    const renderedTasks_books = [];
+    const renderedTasks_restaurants = [];
+    const renderedTasks_products = [];
+    // initialize taskContainers
+    const $taskContainer_movies = $('#movies ul');
+    const $taskContainer_books = $('#books ul');
+    const $taskContainer_restaurants = $('#restaurants ul');
+    const $taskContainer_products = $('#products ul');
+
+    // wrap each task in html and add to corresponding array
+    for (task of tasks) {
+      console.log('task', task);
+      // check which category container to append the task to
+      switch (task.category_id) {
+        case 1:
+          renderedTasks_movies.unshift(createTaskElement(task.input));
+          break;
+        case 2:
+          renderedTasks_books.unshift(createTaskElement(task.input));
+          break;
+        case 3:
+          renderedTasks_restaurants.unshift(createTaskElement(task.input));
+          break;
+        default:
+          renderedTasks_products.unshift(createTaskElement(task.input));
+      }
+    }
+
+    $taskContainer_movies.prepend(renderedTasks_movies.join(''));
+    $taskContainer_books.prepend(renderedTasks_books.join(''));
+    $taskContainer_restaurants.prepend(renderedTasks_restaurants.join(''));
+    $taskContainer_products.prepend(renderedTasks_products.join(''));
+
+    console.log('$taskContainer_movies', $taskContainer_movies[0]);
+  };
+
+  // genereate markup for a single task
+  // task = task (string)
+  const createTaskElement = (task) => {
+    const markup = `
+      <li class='list-group-item'>${task}</li>
+    `;
+    return markup;
+  };
 
 });
